Moved everything in package config
