--- lib/limits_windows.go
@@ -1,9 +1,0 @@
-// +build windows
-
-package lib
-
-// Getrlimit is no-op on Windows, as max fd/process is 2^24 on Wow64
-// Return (16 777 216, nil)
-func Getrlimit() (uint64, error) {
-	return 16_777_216, nil
-}
